+++ docs/v2/introduction/changelog.md
@@ -27,6 +27,7 @@ Pulse.React() is now the simplest way to get started with Pulse, check out the n
 - Services are now found under `pulse.services.service` instead of `pulse.service` potentially breaking
 - Added [mapDataUnderPropName]() config option
 - Added [computedDefault]() config option
+- Added [addStaticData]() as a module function
 - Cleaned up lots of under-the-hood code
 
 ## 2.1.0 - Introducing Modules